Move misc printing functions from app.main to app.utils
